Refactored code and configuration Cleaned subqueries from duplicates
