+++ src/db/yugabyte.py
@@ -5,10 +5,10 @@ import subprocess
 from time import sleep
 from typing import List
 
-from collect_objects import CollectResult, ResultsLoader
+from collect import CollectResult, ResultsLoader
 from config import ConnectionConfig
 from db.postgres import Postgres, PostgresExecutionPlan, PLAN_TREE_CLEANUP, PostgresQuery
-from db_objects import ExecutionPlan
+from objects import ExecutionPlan
 from utils import evaluate_sql
 
 DEFAULT_USERNAME = 'yugabyte'
