+++ src/reports/regression.py
@@ -6,8 +6,8 @@ from sql_formatter.core import format_sql
 from matplotlib import pyplot as plt
 from matplotlib import rcParams
 
-from collect_objects import CollectResult
-from db_objects import Query
+from collect import CollectResult
+from objects import Query
 from reports.abstract import Report
